+++ src/plugins/eth_plugin_handler.c
@@ -1,4 +1,5 @@
 #include <string.h>
+#include "os_lib.h"  // os_lib_call
 #include "eth_plugin_handler.h"
 #include "eth_plugin_internal.h"
 #include "plugin_utils.h"
@@ -305,9 +306,14 @@ eth_plugin_result_t eth_plugin_call(int method, void *parameter) {
     switch (pluginType) {
         case PLUGIN_TYPE_EXTERNAL: {
             uint32_t params[3];
-            params[0] = (uint32_t) alias;
+            // os_lib_call takes a 3-slot unsigned-int array; on device the
+            // pointer size matches uint32_t. Bridge through uintptr_t so
+            // the cast is well-formed on 64-bit host builds too (the
+            // truncation only matters in unit tests where os_lib_call is
+            // stubbed).
+            params[0] = (uint32_t) (uintptr_t) alias;
             params[1] = method;
-            params[2] = (uint32_t) parameter;
+            params[2] = (uint32_t) (uintptr_t) parameter;
             BEGIN_TRY {
                 TRY {
                     os_lib_call(params);
@@ -338,7 +344,15 @@ eth_plugin_result_t eth_plugin_call(int method, void *parameter) {
             // Perform the call
             for (i = 0; i < ARRAYLEN(INTERNAL_ETH_PLUGINS); i++) {
                 if (strcmp(alias, INTERNAL_ETH_PLUGINS[i].alias) == 0) {
-                    ((PluginCall) PIC(INTERNAL_ETH_PLUGINS[i].impl))(method, parameter);
+                    // Two casts through uintptr_t guard against
+                    // -Wpedantic on both ends: the inner one feeds PIC()
+                    // an integer (so its internal (void *) cast is well-
+                    // formed, not function-ptr → object-ptr), the outer
+                    // one turns PIC's void * back into a function pointer
+                    // through uintptr_t (object-ptr → function-ptr is
+                    // equally forbidden by ISO C without the bridge).
+                    ((PluginCall) (uintptr_t) PIC(
+                        (uintptr_t) INTERNAL_ETH_PLUGINS[i].impl))(method, parameter);
                     break;
                 }
             }
